Organizes task logic into submodules
